fix: Delete triggering options preflight

+++ src/pages/secure/api/delete-from-gallery.ts
@@ -1,17 +1,16 @@
 import type { APIRoute } from "astro";
 import { env } from "cloudflare:workers";
 
-export const DELETE: APIRoute = async ({ request }) => {
-	const { id, fileName } = (await request.json()) as {
-		id: string;
-		fileName: string;
-	};
+export const POST: APIRoute = async ({ request }) => {
+	const formData = await request.formData();
+	const id = formData.get("id")?.toString() ?? "";
+	const fileName = formData.get("fileName") ?? "";
 
-	if (!id) {
-		return new Response("Missing id", { status: 400 });
+	if (!id || !fileName) {
+		return new Response("Missing id or fileName", { status: 400 });
 	}
 
-	await env.BUCKET.delete(fileName);
+	await env.BUCKET.delete(fileName as string);
 	await env.DB.prepare("DELETE FROM gallery WHERE id = ?").bind(id).run();
 
 	return new Response(null, { status: 204 });
